+++ semgrep-core/src/analyzing/AST_to_IL.mli
@@ -6,4 +6,7 @@ val function_definition: AST_generic.function_definition
 (*s: signature [[AST_to_IL.stmt]] *)
 val stmt: AST_generic.stmt -> IL.stmt list
 (*e: signature [[AST_to_IL.stmt]] *)
+
+val name_of_entity: AST_generic.entity -> IL.name option
+
 (*e: pfff/lang_GENERIC/analyze/AST_to_IL.mli *)