+++ semgrep-core/src/analyzing/Naming_AST.ml
@@ -345,17 +345,18 @@ let error tok s =
 (* Other Helpers *)
 (*****************************************************************************)
 
-(*s: function [[Naming_AST.is_local_or_global_ctx]] *)
-let is_local_or_global_ctx env lang =
+let is_resolvable_name_ctx env lang =
   match top_context env with
   | AtToplevel | InFunction -> true
   | InClass -> (
       match lang with
       (* true for Java so that we can type class fields *)
-      | Lang.Java -> true
+      | Lang.Java
+      (* true for JS/TS so that we can resolve class methods *)
+      | Lang.Javascript
+      | Lang.Typescript -> true
       | _ -> false
     )
-(*e: function [[Naming_AST.is_local_or_global_ctx]] *)
 
 (*s: function [[Naming_AST.resolved_name_kind]] *)
 let resolved_name_kind env lang =
@@ -367,7 +368,10 @@ let resolved_name_kind env lang =
       (* true for Java so that we can type class fields.
        * alt: use a different scope.class?
       *)
-      | Lang.Java -> EnclosedVar
+      | Lang.Java
+      (* true for JS/TS to resolve class methods. *)
+      | Lang.Javascript
+      | Lang.Typescript -> EnclosedVar
       | _ -> raise Impossible
     )
 (*e: function [[Naming_AST.resolved_name_kind]] *)
@@ -432,7 +436,7 @@ let resolve2 lang prog =
           (* note that some languages such as Python do not have VarDef
            * construct
            * todo? should add those somewhere instead of in_lvalue detection? *)
-          VarDef ({ vinit; vtype }) when is_local_or_global_ctx env lang ->
+          VarDef ({ vinit; vtype }) when is_resolvable_name_ctx env lang ->
             (* Need to visit expressions first so that type is populated *)
             (* If we do var a = 3, then var b = a, we want to propagate the type of a *)
             k x;
@@ -445,6 +449,45 @@ let resolve2 lang prog =
             let resolved_type = Typing.get_resolved_type lang (vinit, vtype) in
             let resolved = { entname = resolved_name_kind env lang, sid; enttype = resolved_type } in add_ident_current_scope id resolved env.names;
             set_resolved env id_info resolved;
+
+        | { name = EN (Id (id, id_info)); _},
+          FuncDef _ when is_resolvable_name_ctx env lang ->
+            (match lang with
+             (* We restrict function-name resolution to JS/TS.
+              *
+              * Note that this causes problems with Python rule/test:
+              *
+              *     semgrep-rules/python/flask/correctness/same-handler-name.yaml
+              *
+              * This rule tries to match two different functions using the same
+              * meta-variable. This works when the function names are not resolved,
+              * but breaks when each function gets a unique sid.
+              *
+              * Function-name resolution is useful for interprocedural analysis,
+              * feature that was requested by JS/TS users, see:
+              *
+              *     https://github.com/returntocorp/semgrep/issues/2787.
+             *)
+             | Lang.Javascript
+             | Lang.Typescript ->
+                 let sid = H.gensym () in
+                 let resolved = untyped_ent(resolved_name_kind env lang, sid) in
+                 (* Previously we tried using add_ident_current_scope, but this
+                  * shadowed imported function names which are added to the
+                  * "imported" scope (globals/block scope is looked up first)
+                  * even when the import statement comes after...
+                  * This broke the following test:
+                  *
+                  *     semgrep-rules/python/django/security/audit/raw-query.py
+                  *
+                  * For now we add function names also to the "imported" scope...
+                  * but do we need a special scope for imported functions?
+                 *)
+                 add_ident_imported_scope id resolved env.names;
+                 set_resolved env id_info resolved
+             | ___else___ -> ()
+            );
+            k x;
 
         | { name = EN (Id (id, id_info)); _}, UseOuterDecl tok ->
             let s = Parse_info.str_of_info tok in
@@ -525,7 +568,7 @@ let resolve2 lang prog =
       );
       V.kpattern = (fun (k, _vout) x ->
         match x with
-        | PatId (id, id_info) when is_local_or_global_ctx env lang ->
+        | PatId (id, id_info) when is_resolvable_name_ctx env lang ->
             (* todo: in Python it does not necessarily introduce
              * a newvar if the ID was already declared before.
              * Also inside a PatAs(PatId x,b), the 'x' is actually
@@ -537,7 +580,7 @@ let resolve2 lang prog =
             add_ident_current_scope id resolved env.names;
             set_resolved env id_info resolved;
             k x
-        | PatVar (_e, Some (id, id_info)) when is_local_or_global_ctx env lang ->
+        | PatVar (_e, Some (id, id_info)) when is_resolvable_name_ctx env lang ->
             (* mostly copy-paste of VarDef code *)
             let sid = H.gensym () in
             let resolved = untyped_ent (resolved_name_kind env lang, sid) in
@@ -566,7 +609,7 @@ let resolve2 lang prog =
           * See: https://golang.org/ref/spec#Short_variable_declarations *)
          | AssignOp(N (Id (id, id_info)), (Eq, tok), e2)
            when lang = Lang.Go && Parse_info.str_of_info tok = ":="
-                && is_local_or_global_ctx env lang ->
+                && is_resolvable_name_ctx env lang ->
              (* Need to visit expressions first so that type is populated *)
              (* If we do var a = 3, then var b = a, we want to propagate the type of a *)
              k x;
@@ -607,7 +650,7 @@ let resolve2 lang prog =
                    (* first use of a variable can be a VarDef in some languages *)
                    (* type propagation not necessary because this does not hold true for Java or Go *)
                    | true, (Lang.Python | Lang.Ruby) (* PHP? *)
-                     when is_local_or_global_ctx env lang ->
+                     when is_resolvable_name_ctx env lang ->
                        (* mostly copy-paste of VarDef code *)
                        let sid = H.gensym () in
                        let resolved = untyped_ent(resolved_name_kind env lang, sid) in