+++ semgrep-core/src/core/ast/AST_generic_helpers.mli
@@ -18,6 +18,10 @@ val funcdef_to_lambda:
   AST_generic.resolved_name option ->
   AST_generic.expr
 
+val name_of_entity:
+  AST_generic.entity ->
+  (AST_generic.ident * AST_generic.id_info) option
+
 val name_of_ids:
   ?name_typeargs:AST_generic.type_arguments option ->
   AST_generic.dotted_ident ->
@@ -46,3 +50,6 @@ val conv_incdec:
 val  conv_class_kind:
   AST_generic_.class_kind * Parse_info.t ->
   AST_generic.class_kind * Parse_info.t
+val  conv_function_kind:
+  AST_generic_.function_kind * Parse_info.t ->
+  AST_generic.function_kind * Parse_info.t