+++ semgrep-core/tests/tainting_rules/ts/basic_crossfun.ts
@@ -1,0 +1,16 @@
+// https://github.com/returntocorp/semgrep/issues/2787
+import React, { Component } from "react";
+class BadPractice extends Component {
+  dangerous() {
+    return location.href;
+  }
+
+  render() {
+    // Testing DOM XSS source --> sink taint tracking in `mode: taint`
+
+    //ERROR:
+    React.createElement("a", { href: this.dangerous() });
+  }
+}
+export default BadPractice;
+